fix : refactoring project structure
