+++ app/types/class-details-subject.ts
@@ -1,0 +1,9 @@
+import { Class } from "./class-details-class";
+
+export type Subject = {
+  id: string;
+  name: string;
+  lecturer: string;
+  semester: number;
+  classes: Class[];
+};
