+++ app/types/class.ts
@@ -1,0 +1,13 @@
+export type Class = {
+  _id: string;
+  subject: string;
+  name: string;
+  quota: number;
+  isFull: boolean;
+  day: string;
+  startAt: string;
+  endAt: string;
+  ruang: string;
+  participants: string[];
+  learningModule: string[];
+};
