+++ app/types/subject-by-semester.ts
@@ -1,0 +1,6 @@
+import { Subject } from "./subject";
+
+export type SubjectBySemester = {
+  _id: string;
+  subjects: Subject[];
+};
