+++ app/types/subject.ts
@@ -1,0 +1,9 @@
+import { Class } from "./class";
+
+export type Subject = {
+  id: string;
+  name: string;
+  lecturer: string;
+  semester: number;
+  classes: Class[];
+};
